EE-631: restructure engine-storage
