+++ execution-engine/engine-storage/src/error/mod.rs
@@ -1,0 +1,4 @@
+pub mod in_memory;
+pub mod lmdb;
+
+pub use self::lmdb::Error;
