+++ execution-engine/engine-storage/src/global_state/in_memory.rs
@@ -2,20 +2,21 @@ use std::collections::HashMap;
 use std::ops::Deref;
 use std::sync::Arc;
 
-use crate::error;
-use crate::global_state::StateReader;
-use crate::global_state::{commit, CommitResult, History};
-use crate::trie::operations::create_hashed_empty_trie;
-use crate::trie::Trie;
-use crate::trie_store::in_memory::{
-    self, InMemoryEnvironment, InMemoryReadTransaction, InMemoryTrieStore,
-};
-use crate::trie_store::operations::{read, write, ReadResult, WriteResult};
-use crate::trie_store::{Transaction, TransactionSource, TrieStore};
 use contract_ffi::key::Key;
 use contract_ffi::value::Value;
 use engine_shared::newtypes::{Blake2bHash, CorrelationId};
 use engine_shared::transform::Transform;
+
+use crate::error::{self, in_memory};
+use crate::global_state::StateReader;
+use crate::global_state::{commit, CommitResult, History};
+use crate::transaction_source::in_memory::{InMemoryEnvironment, InMemoryReadTransaction};
+use crate::transaction_source::{Transaction, TransactionSource};
+use crate::trie::operations::create_hashed_empty_trie;
+use crate::trie::Trie;
+use crate::trie_store::in_memory::InMemoryTrieStore;
+use crate::trie_store::operations::{read, write, ReadResult, WriteResult};
+use crate::trie_store::TrieStore;
 
 /// Represents a "view" of global state at a particular root hash.
 pub struct InMemoryGlobalState {
