+++ execution-engine/engine-storage/src/lib.rs
@@ -1,31 +1,12 @@
 #![feature(never_type, result_map_or_else)]
 
-// core dependencies
-extern crate core;
-
-// third-party dependencies
-#[macro_use]
-extern crate failure;
-extern crate lmdb;
-extern crate parking_lot;
-extern crate wasmi;
-
-// local dependencies
-extern crate contract_ffi;
-extern crate engine_shared;
-
-// third-party dev-dependencies
 #[cfg(test)]
-#[macro_use]
-extern crate lazy_static;
-#[cfg(test)]
-extern crate proptest;
-#[cfg(test)]
-extern crate tempfile;
+use lazy_static::lazy_static;
 
 // modules
 pub mod error;
 pub mod global_state;
+pub mod transaction_source;
 pub mod trie;
 pub mod trie_store;
