+++ execution-engine/engine-storage/src/trie_store/operations/mod.rs
@@ -1,14 +1,15 @@
+#[cfg(test)]
+mod tests;
+
 use std::time::Instant;
 
 use contract_ffi::bytesrepr::{self, ToBytes};
 use engine_shared::logging::{log_duration, log_metric, GAUGE};
 use engine_shared::newtypes::{Blake2bHash, CorrelationId};
 
+use crate::transaction_source::{Readable, Writable};
 use crate::trie::{self, Parents, Pointer, Trie};
-use crate::trie_store::{Readable, TrieStore, Writable};
-
-#[cfg(test)]
-mod tests;
+use crate::trie_store::TrieStore;
 
 const TRIE_STORE_READ_DURATION: &str = "trie_store_read_duration";
 const TRIE_STORE_READ_GETS: &str = "trie_store_read_gets";
